\name{diagnostic}
\alias{diagnostic}
\docType{data}
\title{Frequency of assignment of patients to diagnostic categories}
\description{
Six psychiatrists classified 30 patients into 5 diagnostic categories.
}
\usage{data(diagnostic)}
\format{
  A matrix with 30 rows and 5 columns.
    \describe{
    \item{\code{Depression}}{number of psychiatrists who judged a given patient affected by \dQuote{Depression}}
    \item{\code{Personality disorders}}{number of psychiatrists who judged a given patient affected by  \dQuote{Personality disorders}}
    \item{\code{Schizophrenia}}{number of psychiatrists who judged a given patient affected by  \dQuote{Schizophrenia}}
    \item{\code{Neurosis}}{number of psychiatrists who judged a given patient affected by \dQuote{Neurosis}
  }
  \item{\code{Other}}{number of psychiatrists who judged a given patient affected by \dQuote{Other} diseases
  }
}
}

\references{
Fleiss, J.L. (1971). Measuring nominal scale agreement among many raters. Psychological Bulletin \bold{76, 378-382}
}

\keyword{datasets}
